requirements.txt
numpy>=1.20
nibabel>=5.0
pynrrd>=1.0
